bot/requirements.txt
python-telegram-bot==22.0
sqlalchemy==2.0.30
pymysql==1.1.0
matplotlib==3.8.4